File renames only ("Token -> ColorToken") + update imports
