+++ apps/E2E/src/ColorTokens/consts.ts
