+++ apps/E2E/src/ColorTokens/pages/ColorTokensPageObject.win.ts
