+++ apps/E2E/src/ColorTokens/specs/ColorTokens.spec.macos.ts
@@ -1,5 +1,5 @@
 import NavigateAppPage from '../../common/NavigateAppPage';
-import TokenPageObject from '../pages/TokensPageObject.win';
+import TokenPageObject from '../pages/ColorTokensPageObject.win';
 import { PAGE_TIMEOUT, BOOT_APP_TIMEOUT } from '../../common/consts';
 
 // Before testing begins, allow up to 60 seconds for app to open
@@ -10,7 +10,6 @@ describe('Tokens Testing Initialization', function () {
   });
 
   it('Click and navigate to Tokens test page', async () => {
-    await TokenPageObject.mobileScrollToComponentButton();
     await TokenPageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
 
     /* Click on component button to navigate to test page */
