+++ apps/E2E/src/ColorTokens/specs/ColorTokens.spec.win.ts
@@ -1,5 +1,5 @@
 import NavigateAppPage from '../../common/NavigateAppPage';
-import TokenPageObject from '../pages/TokensPageObject.win';
+import TokenPageObject from '../pages/ColorTokensPageObject.win';
 import { PAGE_TIMEOUT, BOOT_APP_TIMEOUT } from '../../common/consts';
 
 // Before testing begins, allow up to 60 seconds for app to open
