+++ apps/E2E/src/index.consts.ts
@@ -33,4 +33,4 @@ export * from './TabsV1/consts';
 export * from './TextLegacy/consts';
 export * from './TextV1/consts';
 export * from './Theme/consts';
-export * from './Tokens/consts';
+export * from './ColorTokens/consts';
