+++ apps/fluent-tester/src/TestComponents/ColorTokens/index.ts
@@ -1,0 +1,1 @@
+export * from './ColorTokenTest';
