--- apps/fluent-tester/src/TestComponents/Tokens/index.ts
@@ -1,1 +1,0 @@
-export * from './TokenTest';
